Small fixes

+++ src/objects/Home.java
@@ -1,11 +1,12 @@
 package objects;
 
 public class Home {
+
     private String type;
-    private int price;
+    private double price;
     private int m2;
 
-    public Home(String type, int price, int m2) {
+    public Home(String type, double price, int m2) {
         this.type = type;
         this.price = price;
         this.m2 = m2;
